requirements.txt

anthropic==0.34.2
openai==0.28.0
python-decouple==3.8
python-docx==1.1.2
requests==2.32.3
streamlit==1.38.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
groq==0.11.0

